// customer-api/src/utils/enums/RewardEnums.ts
export enum RewardTransactionType {
  EARN_POINTS = 'EARN_POINTS', REDEEM_POINTS = 'REDEEM_POINTS', POINTS_REVERSAL = 'POINTS_REVERSAL',
  EARN_CASHBACK = 'EARN_CASHBACK', USE_CASHBACK = 'USE_CASHBACK', CASHBACK_REVERSAL = 'CASHBACK_REVERSAL', ADMIN_ADJUSTMENT = 'ADMIN_ADJUSTMENT',
}
export enum RewardsGrantedOn { COMPLETED = 'COMPLETED' }